fix(object): 음수 오프셋이 getter 에서 u32 로 새어 개체 드래그가 영구히 죽던 결함
신고(2026-08-15): "자유이동 되지 않고" — 차트를 잡고 끌어도 움직이지 않았다.

원인: 오프셋은 모델에 u32 비트캐스트로 저장되는 부호 있는 값인데, 공용 getter
(common_props_json — 그림·도형·차트 공용)가 unsigned 그대로 내보냈다. 앵커
위/왼쪽 드래그 한 번에 horzOffset 이 4294967295 가 되고, 스튜디오의 다음 드래그
산술(기존값+델타)이 i32 를 벗어나 setter 가 무시 — 이후 그 개체의 이동이 영구히
죽었다. 차트는 기본 배치가 Column/Para 기준이라 위로 끌면 즉시 음수가 되어 가장
먼저 걸렸다(그림은 Paper 기준 양수 영역이라 잠복).

수리: getter 가 vertical/horizontal offset 을 i32 로 캐스팅해 내보낸다.

기존 핀 pin_negative_offset_via_common_setter 는 "음수 오프셋 소실 방지"가 의도
였고 u32 출력은 그 증거 수단이었으므로, 부호 출력 검증으로 갱신했다(같은 의도,
올바른 인코딩). 신규 핀 officex_object_negative_offset: 음수 왕복 + 후속 델타
드래그 재개.

브라우저 검증: 좌상 드래그 → hOff -3000/vOff -4500 로 개체가 실제로 왼쪽·위로
이동, 재드래그 정상(종전엔 재드래그 무반응).

게이트 3533/3533.

+++ src/document_core/commands/object_ops/common.rs
@@ -180,8 +180,12 @@ impl DocumentCore {
             vert_align,
             horz_rel,
             horz_align,
-            c.vertical_offset,
-            c.horizontal_offset,
+            // [2026-08-15 신고 "자유이동 되지 않고"] 오프셋은 u32 비트캐스트로 저장된
+            // 부호 있는 값이다. unsigned 로 내보내면 앵커 위/왼쪽(음수) 드래그 한 번에
+            // 4294967295 가 되고, 스튜디오의 다음 드래그 산술(기존값+델타)이 i32 를
+            // 벗어나 setter 가 거부 — 이후 개체 이동이 영구히 죽는다.
+            c.vertical_offset as i32,
+            c.horizontal_offset as i32,
             text_wrap,
             c.flow_with_text,
             c.allow_overlap,
